refactor : 불필요한 Ajax 요청 삭제, CoverLetterRepository 인터페이스에서 사용하지 않는 메소드 제거 fix : 스크랩 공고가 로그인한 유저의 데이터가 아님에도 접근 가능한 오류 수정
자소서 입력 시, 불필요한 jQuery 요청 삭제로 인한 담당 Controller 삭제, 전체 프로젝트에서 사용하지 않는 코드 제거
자소서 관리와 마찬가지로 취업 공고 스크랩 관리 기능에서도 로그인한 유저의 데이터만 접근하도록 예외처리 추가

+++ src/main/java/Spike/JobDiscription/controller/JobController.java
@@ -46,47 +46,53 @@ public class JobController {
     @PostMapping("/add")
     public String addJob(@ModelAttribute JobDto dto, @SessionAttribute(name = SessionConst.LOGIN_USER, required = false) User loginUser) {
         jobService.create(dto, loginUser);
-
         return "redirect:/jobs";
     }
 
     @GetMapping("/edit/{id}")
-    public String editJobForm(@PathVariable("id") Long id, Model model) {
-        Job job = jobService.showJob(id);
+    public String editJobForm(@PathVariable("id") Long id, Model model, @SessionAttribute(name = SessionConst.LOGIN_USER, required = false) User loginUser) {
 
-        if (job != null) {
-            model.addAttribute("job", job);
-            return "jobs/editJob";
+        if (jobService.isCorrectUser(id, loginUser)) {
+            Job job = jobService.showJob(id);
+            if (job != null) {
+                model.addAttribute("job", job);
+                return "jobs/editJob";
+            }
         }
         return "redirect:/jobs";
     }
 
     @GetMapping("/update/{id}")
-    public String updateJobForm(@PathVariable("id") Long id, Model model) {
-        Job job = jobService.showJob(id);
-        if (job != null) {
-            model.addAttribute("job", job);
-            return "jobs/updateJob";
+    public String updateJobForm(@PathVariable("id") Long id, Model model, @SessionAttribute(name = SessionConst.LOGIN_USER, required = false) User loginUser) {
+        if (jobService.isCorrectUser(id, loginUser)) {
+            Job job = jobService.showJob(id);
+            if (job != null) {
+                model.addAttribute("job", job);
+                return "jobs/updateJob";
+            }
         }
         return "redirect:/jobs";
     }
 
     @PostMapping("/update")
-    public String updateJob(JobDto dto, Model model, UserDto userDto) {
-        Job job = jobService.patch(dto, userDto);
+    public String updateJob(JobDto dto, Model model, UserDto userDto, @SessionAttribute(name = SessionConst.LOGIN_USER, required = false) User loginUser) {
+        if (jobService.isCorrectUser(dto.getId(), loginUser)) {
+            Job job = jobService.patch(dto, userDto);
 
-        if (job != null) {
-            model.addAttribute("job", job);
-            return "redirect:/jobs/edit/" + job.getId();
+            if (job != null) {
+                model.addAttribute("job", job);
+                return "redirect:/jobs/edit/" + job.getId();
+            }
         }
         return "redirect:/jobs";
     }
 
     @PostMapping("/delete")
-    public String deleteJob(Job job) {
-
-        if (jobService.delete(job.getId())) {
-            return "redirect:/jobs";
+    public String deleteJob(Job job, @SessionAttribute(name = SessionConst.LOGIN_USER, required = false) User loginUser) {
+        if (jobService.isCorrectUser(job.getId(), loginUser)) {
+            if (jobService.delete(job.getId())) {
+                return "redirect:/jobs";
+            }
         }
         return "jobs/editJob";
